requirements.txt
asgiref==3.11.1
Django==6.0.3
django-debug-toolbar==6.2.0
djangorestframework==3.16.1
Faker==40.11.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2025.3